updated error handling

+++ PDQ/Epicor Deployment/Remove-Epicor-10.ps1
@@ -2,6 +2,14 @@ $DesktopPath = 'C:\Users\Public\Desktop\Epicor Software\Epicor ERP 10.2 Client'
 
 $InstallPath = 'C:\Epicor\ERP10.2Client'
 
-Remove-Item -Path $DesktopPath -Recurse -Force -Confirm:$false
+if (test-path -path $DesktopPath) {
+    Remove-Item -Path $DesktopPath -Recurse -Force -Confirm:$false
+}else {
+    continue
+}
 
-Remove-Item -Path $InstallPath -Recurse -Force -Confirm:$false
+if (test-path -path $InstallPath) {
+    Remove-Item -Path $InstallPath -Recurse -Force -Confirm:$false
+}else {
+    continue
+}
